Refactor the rest of AndroidUtility.
