+++ app/src/main/java/tech/ula/ui/FilesystemEditFragment.kt
@@ -24,7 +24,7 @@ import kotlinx.android.synthetic.main.frag_filesystem_edit.*
 import tech.ula.MainActivity
 import tech.ula.R
 import tech.ula.model.repositories.UlaDatabase
-import tech.ula.utils.BuildWrapper
+import tech.ula.utils.DeviceArchitecture
 import tech.ula.utils.PermissionHandler
 import tech.ula.utils.CredentialValidator
 import tech.ula.utils.preferences.AppsPreferences
@@ -233,7 +233,7 @@ class FilesystemEditFragment : Fragment() {
             navController.popBackStack()
         } else {
             try {
-                filesystem.archType = BuildWrapper().getArchType()
+                filesystem.archType = DeviceArchitecture().getArchType()
             } catch (err: Exception) {
                 Toast.makeText(activityContext, R.string.no_supported_architecture, Toast.LENGTH_LONG).show()
                 return true
