+++ app/src/main/java/tech/ula/utils/StorageCalculator.kt
@@ -1,0 +1,11 @@
+package tech.ula.utils
+
+import android.os.StatFs
+
+class StorageCalculator(private val statFs: StatFs) {
+    fun getAvailableStorageInMB(): Long {
+        val bytesInMB = 1048576
+        val bytesAvailable = statFs.blockSizeLong * statFs.availableBlocksLong
+        return bytesAvailable / bytesInMB
+    }
+}
